refactoring files in github service
