+++ app/interactors/github.js
@@ -1,6 +1,8 @@
-const { createRepository: create, addDefaultTeamsToRepository } = require('../services/github');
+const { createRepository: create, addDefaultTeamsToRepository } = require('../services/github/repositories');
 
 const createRepository = ({ repositoryName, isPrivate }) =>
-  create({ repositoryName, isPrivate }).then(() => addDefaultTeamsToRepository({ repositoryName }));
+  create({ repositoryName, isPrivate }).then(repository =>
+    addDefaultTeamsToRepository({ repositoryName }).then(() => repository)
+  );
 
 module.exports = { createRepository };
